+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## master (unreleased)
 
+### New features
+
+* [#3](https://github.com/koic/rubocop-faker/issues/3): Make `Faker/DeprecatedArguments` aware of `Faker::Base.unique`. ([@koic][])
+
 ## 0.2.0 (2019-09-04)
 
 ### Bug fixes